bugfixes, working now

+++ config.py
@@ -53,7 +53,8 @@ def compute_weights(n_sm, n_lg, n_half, num_ch):
 
     w_sm = (1./(n_sm * n_sm)) * np.dstack(list(chain.from_iterable(repeat(e, num_ch) for e in [gauss_sm]))).flatten()
     w_lg = (1./(n_lg * n_lg)) * np.dstack(list(chain.from_iterable(repeat(e, num_ch) for e in [gauss_lg]))).flatten()
-    w_half = (1./n_half) * np.dstack(list(chain.from_iterable(repeat(e, num_ch) for e in [gauss_lg]))).flatten()[: n_half * num_ch]
+
+    w_half = (1./n_half) * np.dstack(list(chain.from_iterable(repeat(e, num_ch) for e in [gauss_lg]))).flatten()[: int(n_half * num_ch)]
 
     return np.hstack([w_sm, w_lg, w_sm, w_half])
